Added Models to handle Embedding & Memory Service.
