+++ EmbeddingService/Model/EmbeddingOption/HuggingFaceOptions.cs
@@ -1,0 +1,8 @@
+﻿namespace EmbeddingService.Model.EmbeddingOption;
+
+public class HuggingFaceOptions
+{
+    public string ModelID { get; set; }
+    public string EndPoint { get; set; }
+    public string ApiKey { get; set; }
+}
