--- RAG_API/MemoryFectory.cs
@@ -1,1 +1,0 @@
-﻿
