+++ RAG_API/RAG_API.csproj
@@ -7,12 +7,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="DocX" Version="3.0.0" />
-    <PackageReference Include="iTextSharp" Version="5.5.13.3" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
-  	<PackageReference Include="Microsoft.SemanticKernel" Version="1.9.0" />
-  	<PackageReference Include="System.Drawing.Common" Version="8.0.4" />
-    <PackageReference Include="Microsoft.SemanticKernel.Connectors.HuggingFace" Version="1.9.0-preview" />
+  	<PackageReference Include="Microsoft.SemanticKernel" Version="1.10.0" />
   </ItemGroup>
 
   <ItemGroup>
