--- RAG_API/appsettings.json
@@ -1,42 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "Embedding": {
-    //Choose AzureOpenAI or OpenAI or OLLAMA
-    "EmbeddingType": "OpenAI",
-    "AzureOpenAI": {
-      "APIKey": "Apikey",
-      "EndPoint": "",
-      "Deployment": "",
-      "MaxRetries": 100
-    },
-    "OpenAI": {
-      "APIKey": "ApiKey",
-      "TextModel": ""
-    },
-    "HuggingFace": {
-      "ModelID": "",
-      "Uri": "",
-      "ApiKey": ""
-    },
-    "Ollama": {
-      "endpoint": "http://localhost:3000/api/generate",
-      "model": "phi",
-      "temprature": 0.2,
-      "maxtoken": -1
-    }
-  },
-  "Persistant": {
-    "Type": "Qdrant",
-    "Qdrant": {
-      "KEY": "",
-      "URL": "",
-       "SIZE": 1234
-    }
-  }
-}
